+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Themes: Added rgb variants for the border color tokens
 - Themes: Updated contrast edge cases to be handled with APCA formula
 
+- Autocomplete: Fixed inputAttributes not applying correctly
 - Link: Fixed iOS Safari currentColor
 - View: Updating padding array mentioned in the docs
 - Toast: Added className support
+
+- Calendar: New component